Вроде бы привел оверлей в порядок
