+++ lib/core/misc/date_time_ranges.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:unn_mobile/core/misc/date_time_extensions.dart';
 
 class DateTimeRanges {
   static DateTimeRange currentWeek() {
@@ -44,7 +45,7 @@ class DateTimeRanges {
   }
 
   static DateTimeRange untilEndOfWeek() {
-    final DateTime now = DateTime.now();
+    final DateTime now = DateTime.now().date();
     final DateTime endOfWeek =
         now.add(Duration(days: DateTime.daysPerWeek - now.weekday));
 
@@ -52,7 +53,7 @@ class DateTimeRanges {
   }
 
   static DateTimeRange untilEndOfMonth() {
-    final DateTime now = DateTime.now();
+    final DateTime now = DateTime.now().date();
     final endOfMonth =
         DateTime(now.year, now.month + 1, 1).subtract(const Duration(days: 1));
 
@@ -60,7 +61,7 @@ class DateTimeRanges {
   }
 
   static DateTimeRange untilEndOfSemester() {
-    final now = DateTime.now();
+    final now = DateTime.now().date();
     DateTime endOfSemester;
     if (DateTime.february <= now.month && now.month < DateTime.september) {
       endOfSemester = DateTime(now.year, DateTime.july, 15);
