+++ lib/ui/views/main_page/schedule/schedule.dart
@@ -6,6 +6,7 @@ import 'package:unn_mobile/core/viewmodels/schedule_screen_view_model.dart';
 import 'package:unn_mobile/ui/builders/online_status_builder.dart';
 import 'package:unn_mobile/ui/views/base_view.dart';
 import 'package:unn_mobile/ui/views/main_page/schedule/widgets/schedule_tab.dart';
+import 'package:unn_mobile/ui/widgets/offline_overlay_displayer.dart';
 
 class ScheduleScreenView extends StatefulWidget {
   final int routeIndex;
@@ -43,55 +44,57 @@ class _ScheduleScreenViewState extends State<ScheduleScreenView>
   @override
   Widget build(BuildContext context) {
     final parentScaffold = Scaffold.maybeOf(context);
-    return BaseView<ScheduleScreenViewModel>(
-      model: _viewModel,
-      builder: (context, model, _) {
-        final expanded = _createExpanded(model);
-        return Scaffold(
-          appBar: AppBar(
-            title: const Text('Расписание'),
-            leading: parentScaffold?.hasDrawer ?? false
-                ? IconButton(
-                    onPressed: () {
-                      parentScaffold?.openDrawer();
-                    },
-                    icon: const Icon(Icons.menu),
-                  )
-                : null,
-          ),
-          body: OnlineStatusBuilder(
-            onlineWidget: Column(
-              children: [
-                MediaQuery.withClampedTextScaling(
-                  maxScaleFactor: 1.5,
-                  child: TabBar(
-                    indicatorSize: TabBarIndicatorSize.label,
-                    tabAlignment: TabAlignment.center,
-                    isScrollable: true,
-                    tabs: [
-                      for (final idType in model.tabIdTypes)
-                        Tab(
-                          text: _tabTexts[idType],
-                        ),
-                    ],
-                    controller: _tabController,
-                    onTap: (value) {
-                      model.selectedTab = value;
-                    },
+    return OfflineOverlayDisplayer(
+      child: BaseView<ScheduleScreenViewModel>(
+        model: _viewModel,
+        builder: (context, model, _) {
+          final expanded = _createExpanded(model);
+          return Scaffold(
+            appBar: AppBar(
+              title: const Text('Расписание'),
+              leading: parentScaffold?.hasDrawer ?? false
+                  ? IconButton(
+                      onPressed: () {
+                        parentScaffold?.openDrawer();
+                      },
+                      icon: const Icon(Icons.menu),
+                    )
+                  : null,
+            ),
+            body: OnlineStatusBuilder(
+              onlineWidget: Column(
+                children: [
+                  MediaQuery.withClampedTextScaling(
+                    maxScaleFactor: 1.5,
+                    child: TabBar(
+                      indicatorSize: TabBarIndicatorSize.label,
+                      tabAlignment: TabAlignment.center,
+                      isScrollable: true,
+                      tabs: [
+                        for (final idType in model.tabIdTypes)
+                          Tab(
+                            text: _tabTexts[idType],
+                          ),
+                      ],
+                      controller: _tabController,
+                      onTap: (value) {
+                        model.selectedTab = value;
+                      },
+                    ),
                   ),
-                ),
-                expanded,
-              ],
+                  expanded,
+                ],
+              ),
+              offlineWidget: Column(
+                children: [
+                  expanded,
+                ],
+              ),
             ),
-            offlineWidget: Column(
-              children: [
-                expanded,
-              ],
-            ),
-          ),
-        );
-      },
-      onModelReady: (p0) => p0.init(),
+          );
+        },
+        onModelReady: (p0) => p0.init(),
+      ),
     );
   }
