+++ lib/ui/widgets/offline_overlay_displayer.dart
@@ -1,0 +1,152 @@
+import 'package:flutter/material.dart';
+import 'package:injector/injector.dart';
+import 'package:unn_mobile/core/models/online_status_data.dart';
+import 'package:unn_mobile/core/services/interfaces/authorisation_refresh_service.dart';
+
+class OfflineOverlayDisplayer extends StatefulWidget {
+  final Widget child;
+  final double bottomOffset;
+  const OfflineOverlayDisplayer({
+    super.key,
+    required this.child,
+    this.bottomOffset = 0,
+  });
+
+  @override
+  State<OfflineOverlayDisplayer> createState() =>
+      _OfflineOverlayDisplayerState();
+}
+
+class _OfflineOverlayDisplayerState extends State<OfflineOverlayDisplayer> {
+  final OnlineStatusData onlineStatusData =
+      Injector.appInstance.get<OnlineStatusData>();
+  OverlayEntry? offlineOverlayEntry;
+  @override
+  Widget build(BuildContext context) {
+    return widget.child;
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    onlineStatusData.notifier.addListener(onlineChanged);
+    onlineChanged();
+  }
+
+  void onlineChanged() {
+    WidgetsBinding.instance.endOfFrame.whenComplete(() {
+      if (!onlineStatusData.isOnline) {
+        offlineOverlayEntry = generateOverlay();
+        showOverlay();
+      } else {
+        offlineOverlayEntry?.remove();
+        offlineOverlayEntry?.dispose();
+        offlineOverlayEntry = null;
+      }
+    });
+  }
+
+  void showOverlay() {
+    Overlay.of(context).insert(offlineOverlayEntry!);
+  }
+
+  OverlayEntry generateOverlay() {
+    return OverlayEntry(
+      builder: (context) {
+        return _OfflineOverlay(
+          bottomOffset: widget.bottomOffset,
+        );
+      },
+    );
+  }
+
+  @override
+  void dispose() {
+    onlineStatusData.notifier.removeListener(onlineChanged);
+    offlineOverlayEntry?.remove();
+    offlineOverlayEntry?.dispose();
+    offlineOverlayEntry = null;
+    super.dispose();
+  }
+}
+
+class _OfflineOverlay extends StatefulWidget {
+  final double bottomOffset;
+  const _OfflineOverlay({
+    this.bottomOffset = 0,
+  });
+
+  @override
+  State<_OfflineOverlay> createState() => _OfflineOverlayState();
+}
+
+class _OfflineOverlayState extends State<_OfflineOverlay> {
+  Future<void>? refreshAction;
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: EdgeInsets.only(bottom: widget.bottomOffset),
+      child: Align(
+        alignment: AlignmentDirectional.bottomStart,
+        child: AnimatedContainer(
+          color: const Color(0xFF323232),
+          duration: const Duration(milliseconds: 2000),
+          child: Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Row(
+              children: [
+                Expanded(
+                  child: DefaultTextStyle(
+                    style: Theme.of(context)
+                            .textTheme
+                            .bodyLarge
+                            ?.copyWith(color: Colors.white) ??
+                        const TextStyle(color: Colors.white, fontSize: 14),
+                    child: const Text(
+                      'Нет соединения',
+                    ),
+                  ),
+                ),
+                Padding(
+                  padding: const EdgeInsets.only(
+                    left: 8.0,
+                    right: 12.0,
+                  ),
+                  child: SizedBox(
+                    height: 48,
+                    width: 48,
+                    child: FutureBuilder(
+                      future: refreshAction,
+                      builder: (context, snapshot) {
+                        if (snapshot.connectionState ==
+                            ConnectionState.waiting) {
+                          return const Padding(
+                            padding: EdgeInsets.all(8.0),
+                            child: CircularProgressIndicator(),
+                          );
+                        }
+                        return IconButton(
+                          onPressed: () async {
+                            setState(() {
+                              refreshAction = Injector.appInstance
+                                  .get<AuthorizationRefreshService>()
+                                  .refreshLogin();
+                            });
+                          },
+                          icon: Icon(
+                            Icons.refresh,
+                            color: Theme.of(context).primaryColor,
+                          ),
+                        );
+                      },
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
